Allow for a valid queue to appear by default in ticket-creation form
Previously, the code was inserting a choice of all dashes (with an
empty-string queue ID) as the first choice for Queue in the dropdown
when a user is creating a new ticket. This puts some burden on the
user by forcing them to learn what a queue is and select the right
one.

This commit drops the invalid queue choice, making the form a bit
easier to use.