Bug fix, change ticket.customer_id to ticket.submitter_email.
The Ticket model does not contain a 'customer_id' field. For non-staff
users who have permission to view their own tickets, this conditional
statement causes an error since 'ticket.customer_id' is undefined.

The Ticket model does contain a submitter_email field which is used
instead. If the current User is the submitter of the ticket in question,
then permission is granted for that user to view the ticket.