Hide certain fields on the create_ticket form for non-staff users.
Non-staff users should see fewer fields on the 'Submit A Ticket'
form compared to admin users. These fields, though not removed
from the form, are hidden in this commit to preven the user from
editing them.

Some fields, such as 'submitter_email', need to be included even
if we do not want users editing them. The submitter email field
is used in determining which tickets were submitted by which
user (and thus, used to populate the 'Your tickets' list on the
dashboard).