Pass global_tickets list to the helpdesk dashboard.
A new object named 'global_tickets' is passed to the context
when rendering the dashboard page.

This object contains a list of all tickets with 'viewable_globally'
set to True. In addition, to allow non-staff users to view global
tickets, a new check is added to '_is_my_ticket' which grants
permission on all global tickets.