Remove text in emails which instruct users to reply to the email.
When tickets are updated or marked as 'Resolved', the email which
is sent to the user contains text instructing the users to reply
to the email if they need to provide additional information when
tickets are updated, or if they feel the resolution is not adequate.

In this commit, we remove that text from the email template, and
instead instruct users to re-open the ticket if the resolution
is not adequate, or view the ticket online to provide additional
information.