Remove text in emails informing users to reply to the email.
When new tickets are created, the courtesy email informing users
their tickets have been received previously contained a line
informing users to reply to the email using the ticket number in
the subject.

In this commit, that line is removed from the email. Instead, users
are given a link to the ticket on the site and asked to use that
to provide further information or view any updates.