Handle new user creation in email_address_update signal
Because this is a pre_save signal, if a user is about to be created,
there is no user to query for and an exception is thrown. Here we change
the get() to a filter() and check to see if a user was returned. If no
user is found, then we can just return because they would not have any
tickets to update.